Upgrade unattended_reboot module
The unattended reboot module added MAILTO="" to our standard cron jobs,
which meant anything below it was eaten by this, and consequently we
lost our logs.

The module has been updated to use /etc/cron.d rather than being added
to a user crontab. This updates the module we use to push this change.