Stop search benchmark jobs from notifying Slack
Ideally these Jenkins jobs would notify the #search-team Slack channel,
but they actually spam the #2ndline channel and we haven't been able to
work out how to configure the Jenkins Job Builder to send the
notification to the right place.

We are planning to change how the search benchmark runs anyway, so the
simplest solution for now is to just switch off the Slack notifications.
They will still notify Icinga if they fail completely.