Add jenkins job to run analysis on nav links
The new job runs a monthly analysis of links displayed in the new navigation
pages. The output of this is needed for Performance Analysts to produce reports
on how the new navigation is performing.

The rake tasks needs a credentials JSON file in the root path of the
repository, which will be built and placed in the correct place with
credentials stored in hieradata.