Update nav monitoring jenkins jobs
Since the job name is different than the repo name, we need to `cd` into
the directory before we can run bundler. Furthermore, we need to copy
over the credentials file for google analytics into the repository after
it's cloned in order for the rake tasks to work as expected.

This fixes the following issue in the tagging monitor job:

```
rbenv: bundle: command not found

The `bundle' command exists in these Ruby versions:
```

Source:
https://deploy.integration.publishing.service.gov.uk/job/govuk_tagging_monitor/17/console

Now that the directory exists, we will be able to fix this too:

```
Could not set 'present' on ensure: No such file or directory -
/var/lib/jenkins/govuk_navigation_link_analysis/govuk-tagging-monitor-2f614b9b92c2.json
```

When the file is available, the following error will be fixed:

```
Errno::ENOENT: No such file or directory @ rb_sysopen - govuk-tagging-monitor-2f614b9b92c2.json
```

Source:
https://deploy.integration.publishing.service.gov.uk/job/govuk_navigation_link_analysis/1/console