Add second transition stats key
You can only use a deploy key once on Github, which means that a second
one needs to be added. This was the behaviour of the original Puppet
code, but I incorrectly thought it wasn't required.

The script would not have worked without adding the SSH config either as
when it clones the repo for the first time it references the host in
that config.