Update jenkins-cli script
The latest version of jenkins-cli.jar changed the way in which SSH
identity files are used, which caused the following error to be thrown:

`ERROR: anonymous is missing the Overall/Administer permission`

Instead of using the SSH identity option which now explicitly requires
enabling an SSH server in Jenkins itself, auth using the API user and
token that we already use for Jenkins Job Builder.

Instead of referencing this in two places in hiera, pass in the Jenkins
api user parameter using the main class to the two sub classes that use
it.