Rearrange disk ordering on asset-slave-1.integration
After rebooting, the disks are now in a different order and Puppet is failing. In the short term, this will fix Puppet.
What would be better is to uniquely identify the underlying 'disk' and assign the PVs to those but I'm not sure this is possible in a VMWare environment