Ensure /mnt/logs_cdn is writeable by group
Ensure that the logs_processor user is able to to write to the directory
that holds CDN logs. This is so it can touch a file called "done" when it has
processed a log.

We also need to amend logrotate so that it is able to rotate logs.
Logrotate will complain if it's rotating logs in a directory that has
insecure permissions where a user is not specified for the rotation.
Update the logrotate configuration to ensure the logrotation still
occurs, which is required by the logs processing script.