Ensure transition logs cache files are rotated
The transition logs processing script uses the cache directory to add
data for files which have been processed. These can quickly use up disk
space, so add logrotation on these files.

We use the "olddir" directive to move old cache files into the "archive"
directory. This is relative to the log directory as per the man page for
logrotate.