Fix invalid Jenkins job config
Attempts to update the Jenkins jobs fails with the output

```
childProjects.text = data['project']
TypeError: list indices must be integers, not str
```

It looks like because it’s only a single value it should just be a hash rather than an array of hashes - other trigger definitions in the other config files do not have this leading dash either.

YAML is hard.