Update Puppetdb configuration for version 2
With version 2.0.0 we still need to run puppetdb-ssl-setup to configure jetty.ini
properly and create the ssl files, but this version is not using the same certificate
store than 1.x so we need to add a variable for that.

With version 2.0.0, manage the JAVA_ARGS configuration in the default service
options file.