Add a second hiera structure, for use by AWS
This seperates the existing puppet hiera from the new AWS values
and makes changes less risky to the existing environments while
also allowing us to remove the new settings easily if we need to.

I've included the current common and integraiton config files, as
we've been testing with those, but with a change to app_domain
to derive it from the stack tag passed in to puppet in the AWS
terraform code. This is an attempt to remove most needs to have
per stack versions of the config file.