Use clamd to scan files
Currently, for every file we scan for viruses, we boot a new instance
of `clamscan`. This means it has to load the entire virus definition
database each time, and each file takes a minimum of 11 seconds to scan.

`clamd` is a server version of `clamscan`. It keeps the definition
database in memory and has the file passed to it from `clamdscan`,
which acts similarly to `clamscan`. We’ve tried this approach before
but we didn’t use `--fdpass`. This made things fail as `clamd` couldn’t
access the files we were trying to scan. Specifying `--fdpass` passes
the file descriptor from `clamdscan` (which does have access) to
`clamd`.

This should bring the per file scan time down from ~11-14 seconds to
~1-2 seconds.