Move parameters to configurable parameters
Previously these were defined by if an instance was Jenkins v1 or v2.

Since all instances should now be v2, these can move into being
configurable parameters.

Specifically the CSRF parameter is configurable as it meant that some of
our jobs that would normally be triggered by another Jenkins instance
failed. For the time-being, this can be set to false, and when we've
resolved the issue it should be set to true as default.