Add S3 bucket config to Asset Manager
We're enhancing Asset Manager to upload files to, and serve files from
S3. This PR sets the AWS environment variables required by Asset
Manager.

We're safe to use the standard AWS environment variable names because we
rely on `govuk_setenv` to provide each application with its own
environment.

See the related PRs in asset-manager and govuk-terraform-provisioning:

* https://github.com/alphagov/asset-manager/pull/74
* https://github.com/alphagov/govuk-terraform-provisioning/pull/125