Add API user
Jenkins Job builder requires a user to authenticate, as does the
jenkins-cli stuff. I couldn't find where the ci_alphagov user was
created anywhere in Puppet, so take this opportunity to change the name
of the api user that is used for auth. I suspect the user had been added
manually in the past.