Begin basic node class for s_db_admin
We're looking to provision our databases in RDS rather than provision
EC2 machines with databases on them.

To create, administer and backup databases using Puppet, we need something to
use that can do this in our current Puppet estate.

For now just install packages to used to work with our current
versions MySQL and PostgreSQL.