Add conditional to avoid govuk_host related errors
There is a class which purges govuk hosts which is based upon a class
that we specifically exclude when running in AWS.

We have a requirement to run a "db-admin-1" machine in AWS only to
manage our RDS instances.

This commit adds a whitelist array, and then if a machine exists in this
whitelist then it will not run the check that makes the tests fail.