Add support for multiple queue/routing key combinations
This allows a application to connect with either one or two
queues where previously this limit for a single queue.

As it is not possible to update the user permissions to
accesses an additional queue when it is created we are
creating a new user which will then be able to connect to both
queues.