Add ssh-proxy command to jumpboxes in AWS
The usage would be:

`ssh-proxy delana jenkins`

This looks up the instance IP address using tags so you can jump
straight to it from the jumpbox. It includes the stackname so that it's
a generic script that can be called from a client machine on different
stacks.

If more than one host exists in a role (ie in an auto-scaling group),
then it will select the first one.