Change notification period of sync-asset-manager-from-master icinga check
In integration, the data sync in the morning copies data to the asset
master, but the sync between the asset-master and the slaves usually
does not finish until midday. This means that there are always alerts in
integration in the morning before the sync to the slaves has had a
chance to run.