Add later version of terraform to CI machines
Currently we use an old version of terraform (0.8.1) for a few things in our
stack:

 - Deploying govuk-dns on Deploy Jenkins
 - Testing govuk-dns on CI
 - Deploying code in govuk-terraform-provisioning on Deploy Jenkins

We're currently migrating the GOV.UK stack to AWS, and use a much later
version of Terraform to do this (0.9.10). We should be testing the code
we're writing, and so we should add a later version of Terraform on a
couple of CI agents to allow this.