Update elasticsearch config for machines in AWS
Instances in AWS will do a discovery for the cluster using tags and the
aws-cloud plugin (which we already installed for our current versions of
Elasticsearch).

Update the configuration so that if an instance is in AWS, it will apply
the correct config.

The cluster_name is set as a default which defaults to use a tag based
upon the stack name. This is purely for convenience.