Temporary fix while we're migrating things
I've found pretty often that hosts are unable to do looks up for the
Puppetmaster and so Puppet isn't running consistently across all
machines.

The problem is that resolv.conf is being overwritten and the search
domains are being lost.

Currently I don't want to dip all my time into resolving this issue, so
I propose adding a cronjob which restarts things if it can't resolve the
puppet host.