Don't set Elasticsearch local_proxy for AWS
We use the class govuk_elasticsearch::local_proxy to set up an nginx
vhost on the local application server which load balances request to our
elasticsearch cluster.

In Amazon our cluster will sit behind an ELB, so as far as I can gather
this isn't required as we can just set the applications to use the DNS
name tied to the ELB.

As it currently stands, including this means that Puppet is failing to
run on the machines where it is included.