Update Publishing API Metrics grafana board
This is mostly pulled by exporting the Grafana JSON which has
unfortunately created a bigger diff than would be appreciated.

This splits the Publishing API request speed graphs into two, one for
reads the other for writes; includes endpoints that were missing; and
fixes ones that no longer worked.

The errors graph has been simplified to show All errors, GET request
errors and POST/PUT/PATCH/DELETE errors.

The memory metrics have been updated for the Publishing API changing
boxes.

At time of writing the DB size doesn't seem to be working but I believe
that is because of an issue with statsd on the Postgres host rather than
this script.