Update govuk_node_list for AWS
In AWS, we can find out the hostnames for our instances using tags. We
still require PuppetDB to discover hosts which are assigned with
specific classes, but we can update the functionality for listing
specific node classes by using boto.

This has been rewritten in a way that should not be visible to users or
scripts so can be an in-place drop in. This should ensure we're able to
continue using our current deployment scripts without impact.