Add db_admin config for PostgreSQL
We wish to manage a PostgreSQL instance in Amazon RDS, and we wish to do
it using the current module we use to manage all our databases, users
and permissions.

The module is a little restrictive in how we manage it, and it requires
installing a local PostgreSQL server. We then set the connection
settings which point to our relevant RDS instance.