Do not restart Mysql on config change
The latest puppetlabs-mysql upgrade is going to introduce some
minor changes in the Mysql configuration. This change updates our
govuk_mysql module to not restart the servicemif the configuration
is updated, so we can deploy the previous changes on Production.

After deploying Puppet on Production, we can revert this change if
we want to restore the previous behaviour.