Limit process graph to last 15m
This panel is supposed to show you when your deploy is finished (i.e.
new processes have been started and the old ones killed).

Seeing process counts for the whole time range of the dashboard is
counterproductive.

This is the same as 36847fb1, just with the change applied to the
deployment dashboard, rather than the processes dashboard.