Allow applications to set the SENTRY_DSN env var
The `SENTRY_DSN` env var is used by the Sentry client to report
exceptions. This commit allows us to set the var for applications. To
use it we'll have to pass in the `sentry_dsn` variable via hieradata
into `govuk::app`. The hieradata will be set in govuk-secrets.

The DSN for collections & content-tagger are added here because we'll
be testing with them first. Once the test is successfull we'll add the
other applications as well.