Ensure permissions are set correctly for RDS
We're looking to manage our PostgreSQL databases and roles in Amazon RDS
using the current Puppet module where possible.

When you create a new RDS instance you are assigned an "rds_superuser".
This user has most privileges, but it does not have the same level of
privileges that we have when managing a locally installed PostgreSQL
server.

This means that when we call the postgresql::server::db defined type
(which creates a role and then a database which is owned by that role)
it rds_superuser does not have the correct level of privileges.

To allow the rds_superuser to create what we want, we must ensure that
it has membership of the created role that owns the database we're
trying to create.

We export the settings when we call the defined type in the class, so we
must also export the command that assigns this membership to it's own
defined type, and collect it in the db admin class.

We have to pass in a conditional (boolean "rds") to ensure this is only
run when specified in an RDS instance.

We must also pass the "rds_root_user", and since it's a defined type it
cannot be set in hieradata. I have set it as the current testing default
we have, and in the future we can make this a generic admin role that
matches what we set in Terraform when building the RDS instances.

This does work, but I have noted that each time Puppet runs we get
output notices about an ALTER ROLE on each user. I am aware of this
issue and will look to fix it in the future.