AWS puppetserver support
On AWS we are going to run the puppet master service from puppetserver
instead of Nginx + Unicorn, to help the transition to Puppet 4 and
support certificate extensions for node classification.

This requires a new puppetserver manifest and some adjustments to
install Openjdk and remove conflicting gems that are installed during
the Puppet master bootstrap. Nginx is still installed to enable access to
Puppetdb, but we should review this.