Run router as container in AWS
Router is pretty integral to running the frontend apps. In AWS, we do
not have access to our CI environment, so we need to make a bunch of
changes to allow us to push artefacts to S3.

We already have router in docker hub and it gets pushed there
automatically. There aren't too many changes and it actually removes a
bit of code, so run router as a container in AWS.