Remove error log output
By default a container will log to syslog (as per our govuk_docker class). The
env var in this case is related to inside the container, so remove. The
default for the app is to error to STDERR.

If we want to take this to Production we may want to change this to go
to a specific file.