Configure asset-manager to proxy to S3
This config allows the asset-manager Rails app to instruct nginx to
proxy requests for assets to S3. This is going to allow us to
performance test this approach.

The `cloud-storage-proxy` location is marked as `internal` which means
that it can't be accessed from outside nginx. We use this location by
setting the `X-Accel-Redirect` header to
`/cloud-storage-proxy/<url-of-asset-on-s3>` in our Rails app.

We're consciously not enabling caching of the assets from S3 at the
moment. We want to test the performance without it before deciding
whether it's necessary to add it.

I was hoping to be able to use a regexp/wildcard to hide all `x-amz-*`
headers but it doesn't seem to be possible.

We're explicitly setting the S3 Host (using `proxy_set_header Host
$proxy_host;`) in order to 'reset' the HTTP headers set in the outer
configuration block (e.g. `X-Accel-Mapping`). The comment attemts to
explain the reason for this.