AWS govuk_node_class prevent spoofing
Update `govuk_node_class()` to obtain the class_name on AWS from the
`$::trusted` hash variable instead of the `$::aws_migration` fact in
order to prevent spoofing.

On AWS environments, the certificate request has been updated to
include the following extensions, that will appear in the `$::trusted['extensions']`
hash when we run `puppet agent` on the clients:
 - pp_instance_id: AWS instance Id
 - pp_image_name: AWS image name
 - 1.3.6.1.4.1.34380.1.1.13: Instance role, populated at provisioning
time with the value of the `aws_migration` tag. This OID should resolve
to `pp_role` in Puppet 4.1
 - 1.3.6.1.4.1.34380.1.1.18: AWS instance region. This OID should resolve
to `pp_region` in Puppet 4.3

When we run `puppet agent`, `govuk_node_class()` will return the content of
`$::trusted['extensions']['1.3.6.1.4.1.34380.1.1.13']` if it's not empty or null.
Otherwise the previous `$::trusted['certname']` is used to extract the class_name.

When we run `puppet apply --trusted_node_data`, `govuk_node_class()` will return the
value of the fact `$::aws_migration` if it's not empty or null. Otherwise
`$::trusted['certname']` will be used.