AWS add puppet node clean cron job
Add a cron job to remove nodes from Puppetdb that no longer exist on the
AWS environment. We should refactor the AutoScalingGroups to use queue
notifications, but for now this will help preventing errors on monitoring
and cache servers related with duplicated stored resources that are using
`$::fqdn` as resource key.