Email users upstream of a failing Smokey build
Prior to this change, Smokey would only notify a distribution list on
failure. If people aren’t watching that list, they will not be notified
of failures.

This change still notifies the distribution list, but on top of that it
tries to work out who kicked off the job upstream of Smokey causing it
to fail, and emails them.

In other words, if someone kicks off one of:
- Deploy App
- Deploy Puppet
- Deploy Licensify
- Network Config Deploy

...and then Smokey fails, then they will be notified by email, if that
email can be inferred.

Additionally, if changes were made to Smokey, and then it fails, then
the “culprits” who broke the build will also be notified.

In order to achieve this behaviour, we need to use the Jenkins
[`email-ext`](https://wiki.jenkins.io/display/JENKINS/Email-ext+plugin)
plugin.

This plugin offers far more customisation than the plainer `mailer`
plugin, and - importantly - allows us to email addresses that are
unknown to Jenkins. After Jenkins’ security advisory
[`SECURITY-372`](https://jenkins.io/security/advisory/2017-03-20/),
it was not possible to email addresses not associated with Jenkins users.

However, since we authenticate via Github, Jenkins does not actually
consider our email addresses to be associated with our projects.
`Email-ext` allows us to overcome this by running Jenkins with the
`-Dhudson.tasks.MailSender.SEND_TO_USERS_WITHOUT_READ=true` `JAVA_ARG`,
which is also added in this change.