Add Aptly logstash repo
At the moment we are downloding a very old version of logstash from a site that
has an expired certificate, so the download is failing on new installations.

We are going to package the Logstash jar and uploaded to an internal repository until
we upgrade to the latest version. This change creates the Aptly repository to store
the package.