s_logging logstash package
At the moment we are downloding a very old version of Logstash from a site that
has an expired certificate, so the download is failing on new installations.

The Logstash jar has been packaged and uploaded to an internal repository, so
this commit is updating the s_logging class to get the jar file from the package
instead of downloading it. This should work for both VCloud and AWS. The package
is just going to put the jar file in the same location as the previous `curl`
resource, so the rest of the class remains as it was.

s_logging will need to be refactored when we move to the elastic logstash Puppet module
and upgrade the Logstash version.