Nginx limit rate alert on cache servers
At the moment we monitor limit rate errors (HTTP 429) on the cache services
(www.govuk.uk and assets) only in Grafana. This change creates a Icinga alert,
similar to 5xx errors, to monitor on HTTP response code 429 during office
hours.

We should monitor the alert to check that the limits and metric window are
correct before enabling out of office hours.