Add an uptime collecting service
This sends the uptime data it collects to statsd which allows it
to be viewed in Graphite.

We decided to go down the route of a service rather than taking
advantage of the Icinga checks because we wanted to have control
over when the /healthcheck endpoint will be polled.

At the moment this only works with content-store and
publishing-api but it should be possible to make it work with
more apps in the future.