Use ETag & Last-Modified from Asset Manager when proxying to S3
This is a follow-up to the PR in which we made it possible to configure the
Asset Manager to proxy asset requests to S3 via Nginx [1]. Since then we've
merged an Asset Manager PR [2] which allows us to divert a percentage of asset
request traffic to be proxied to S3. We plan to roll this out gradually while
carefully monitoring the load on and performance of the production stack.

Unfortunately Nginx and S3 use different algorithms to generate an ETag
response header and, since the S3 copy of each asset has a different last
modified time, the Last-Modified response header values are different too.
So in order to avoid unnecessary cache invalidation and hence confound our
monitoring of load & performance, we've made a change [3] to the Asset
Manager Rails app to generate Nginx-esque ETag & Last-Modified response
headers.

The changes to the Asset Manager Nginx config in this commit store the response
header values from Rails and then add them to the response from S3 before the
response is sent back to the user thus avoiding the unnecessary cache
invalidation mentioned above.

Note that this is only an intermediate step; eventually we're planning to
*redirect* asset requests to S3 and at that point we won't have control over
these response headers. However, at that stage the load & performance concerns
will transfer to S3, which is much less of a concern.

[1]: https://github.com/alphagov/govuk-puppet/pull/6280
[2]: https://github.com/alphagov/asset-manager/pull/152
[3]: https://github.com/alphagov/asset-manager/pull/157