Add resolver to nginx config in AWS
On startup, nginx does DNS lookups for upstream hosts and caches the
result. In AWS, when ELBs scale up or down, they are liable to change
endpoints. We use DNS to reach the appropriate endpoints, but unless we
restart nginx then upstream requests will timeout.

By adding a resolver directive, it means that if the upstream request
times out, it will do a DNS search:

https://nginx.org/en/docs/http/ngx_http_core_module.html#resolver

The resolver value is for the VPC nameserver, which is always the same
in a VPC based upon the network range:

http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_DHCP_Options.html#AmazonDNS