Add app_domain_internal to nginx in AWS
In AWS, we're using a combination of external and internal domain names.
For places where we only want the internal domain name used, it's easier
to add both the internal and external app domains in the nginx vhosts.

This adds some (horrible) logic that strips the magically assigned
appended domain to the title of the defined type, and assigns the
internal domain name.

It then adds this as an alias to the server_name section of the nginx
template.