Escape commas for Emergency Banner Deploy task
Rake is not very good at handling arguments with commas in. It has
surprising behaviour, such as:

`rake test[‘foo, bar’]` will give you two arguments: `foo` and `bar`.

If you escape the comma, then they are provided as a single argument
with a comma: `foo, bar`.

This change escapes the commas in the provided arguments, in order
to reduce human error when manually escaping. Since the command is
executed inside speech marks, this change also escapes quotes.