Install oracle7 similar to oracle8
Currently we install jdk7 using a tarball in an S3 bucket the location
of which I can't find.

This changes it to use the install method used for oracle8 by making the
class more open and passing in a parameter.

We use this class for two things:

 - Installing a really old version of logstash
 - Licensing

I'm happy to change the method simply to install logstash, which we will
hopefully upgrade soon anyway.