Increase Postgres Primary disk space
Postgres Primary is running out of space when performing backups in the
production environment since there is not enough disk space available
for the uncompressed backups.

At current time we have a 126G partition on integration and production
and a 156G one on staging. So this commit only adds extra drives to the
integration and production environments.