Make search benchmark alerts non-critical
Downgrade the alert level from the default value of 2 (critical) to 1
(warning).

Search benchmark failures do not need to be investigated urgently. They
just report on the current state of search results, and there are other
services which monitor the availability of search which will alert us
to more serious problems.