Simplify consumer class by removing exchanges
RabbitMQ exchanges are only relevant to the publisher of the messages, so its a
bit weird for the consumer stuff to create these.

They're only used in one place for creating test infrastructure.

Simplifying the consumer makes it easier to change the search infrastructure.
At the moment this is difficult because there is a lot of logic for
constructing permission strings.

The ultimate goal is to separate queue creation so we can set up multiple
queues per consumer.