Remove permission logic from consumer class
This should make it easier to audit what permissions we are giving each user.

The current values are a bit surprising, as now that puppet manages the
creation of queues, the consumers themselves don't need full
write/configuration permission.

I'm keeping this the same for now, because my goal is to separate out the queue
creation, so we can have multiple queues and bindings for rummager.