collect prospectors in the filebeat class
We only want filebeat::prospector instances to have an effect on nodes
which actually have filebeat installed.

Previously, this was achieved in nginx::log by having an `if`
statement conditional on $::domain.  However, I don't really want to
duplicate that conditional logic everywhere we might want to have a
prospector.

This commit makes filebeat::prospector create virtual file resources,
which only get realized in the filebeat class.  This means we can
remove the conditional logic from where we declare prospectors.

(For this, I take advantage of puppet's automatic tagging: the file
created in filebeat::prospector is automatically tagged with
filebeat::prospector).

We use virtual resources elsewhere with things like @collectd::plugin,
so there is prior art for this kind of thing.

I wasn't sure whether to make the `file` resource virtual, or the
`filebeat::prospector` resource.