Log slow Elasticsearch queries
Some queries from finder-frontend to Rummager have been timing out
recently, which is causing 500 errors on finders. This would be easier
to investigate if we could see which Elasticsearch queries were running
slowly and at what times.

Only enable slow query logging on the rummager-elasticsearch boxes for
now. We can easily enable it later on logs-elasticsearch or in
development if it would be useful.

https://trello.com/c/saDruTlG/307-investigate-1-day-intermittent-rummager-timeouts