Add Rummager  workers to deployment dashboards
The Rummager workers have a different namespacing to many of the other
applications, such as `Indexer` and `GovukIndex`. Unfortunately it does
not look like Graphite supports wildcard matching across the `.` that
separates path segments. Any workers that do not exist will simply not
appear so the dashboard will remain uncluttered.