Try to slow down SharePoint crawlers causing too many 429s
We've identified that this user agent that is used by SharePoint servers
(https://social.msdn.microsoft.com/Forums/en-US/3ea268ed-58a6-4166-ab40-d3f4fc55fef4)
is hitting GOV.UK with too many requests causing our mointoring to go above it's threshold
often.

We're adding a crawl-delay here in the hopes it'll slow down these requests but we're
not sure if the crawl-delay directive is implemented by SharePoint, so this may need to be reviewed later.