Add Nginx config for Asset Manager Rails asset pipeline assets
We have enabled the Rails asset pipeline for Asset Manager in
alphagov/asset-manager#199 and enabled asset precompilation in
alphagov/govuk-app-deployment#183. In order to make the Rails asset
pipeline assets in Asset Manager available to members of the public, we
need to add this Nginx configuration which conforms to the same pattern
as other Rails apps, e.g. calculators.

As explained in alphagov/asset-manager#199, we need this in order to
move the `thumbnail-placeholder.png` image from Whitehall to Asset
Manager as part of the plan to serve Whitehall assets from Asset Manager
as described in alphagov/asset-manager#200.

I've run `govuk_puppet` with these changes on my development-vm and it
added the following lines to
`/etc/nginx/sites-available/assets-origin.dev.gov.uk` as I expected:

    location /asset-manager {
      proxy_set_header Host asset-manager.dev.gov.uk;
      proxy_pass http://asset-manager.dev.gov.uk;
    }

I also then tested the change by adding an image to Asset Manager at
`app/assets/images/example.png`, using the Rails console to obtain the
URL path for the image, and then requesting the image using `curl` via
the `assets-origin.dev.gov.uk` host. The Asset Manager Rails app
successfully served the image at the following URL, so I'm happy that
this configuration change is correct.

    http://assets-origin.dev.gov.uk/asset-manager/example-721110d114be906508233916f3750f05576596ef0d51ed839c931912ca67eede.png

Closes alphagov/asset-manager#202.