Add check for filebeat
In our code we set a conditional that either sets `true` or `undef`:

https://github.com/alphagov/govuk-puppet/blob/a4d60bd1f1cd9782658b5d7e65032e01ca6c1870/modules/govuk/manifests/app.pp#L361-L366

The Ruby code in the template treats both as a string, so are set to
`true`, but if `undef` is set, then this is written into the
configuration file, which causes filebeat to fail on start up:

```
2017-09-19T12:09:43Z CRIT Exiting: Error in initing prospector: can not convert 'string' into 'bool' accessing 'filebeat.prospectors.0.json.add_error_key' (source:'/etc/filebeat/conf.d/asset-manager-app-err.yml')
Exiting: Error in initing prospector: can not convert 'string' into 'bool' accessing 'filebeat.prospectors.0.json.add_error_key' (source:'/etc/filebeat/conf.d/asset-manager-app-err.yml')
```

This adds a condition that if the value of the key "add_error_key" is
set to a boolean then print it, otherwise do not add that option.

This also adds some tests to help catch any errors.